+++ WebApp/Models/SEG/UsuarioPermiso.cs
@@ -9,14 +9,14 @@ namespace Core.Models.SEG
     [Table("SEG.UsuariosPermisos")]
     public class UsuarioPermiso
     {
-        public UsuarioPermiso(int usuarioId, int permisoId)
+        public UsuarioPermiso(long usuarioId, long permisoId)
         {
             UsuarioId = usuarioId;
             PermisoId = permisoId;
         }
         [Required(ErrorMessage = "El campo Id es obligatorio.")]
-        public int Id { get; set; }
-        public int? UsuarioId { get; set; }
-        public int? PermisoId { get; set; }
+        public long Id { get; set; }
+        public long? UsuarioId { get; set; }
+        public long? PermisoId { get; set; }
     }
 }
